requirements.txt
paho-mqtt
numpy
matplotlib